+++ src/index.ts
@@ -1,2 +1,8 @@
 export type { PluginI18nextExtractorOptions } from './options.js';
+export type {
+  AfterExtractPayload,
+  I18nextExtractorWebpackPluginHooks,
+  RenderExtractedTranslationsPayload,
+} from './hooks.js';
+export { getI18nextExtractorWebpackPluginHooks } from './hooks.js';
 export * from './pluginI18nextExtractor.js';